cardano-cli:  revert to pre Cardano.CLI.Helpers
