+++ cardano-node/cardano-node.cabal
@@ -15,7 +15,6 @@ library
   hs-source-dirs:      src
 
   exposed-modules:     Cardano.Chairman
-                       Cardano.CLI.Helpers
                        Cardano.CLI.Ops
                        Cardano.CLI.Run
                        Cardano.Common.Protocol
