+++ cardano-node/src/Cardano/CLI/Run.hs
@@ -41,35 +41,63 @@ module Cardano.CLI.Run (
   , NewTxFile(..)
   ) where
 
-import           Prelude (show)
+import           Prelude (String, error, show)
 import           Cardano.Prelude hiding (option, show, trace)
 import           Test.Cardano.Prelude (canonicalDecodePretty)
 
 import           Codec.Serialise (serialise, deserialiseOrFail)
 import           Control.Tracer
 import           Data.Bits (shiftL)
+import qualified Data.ByteArray as BA
+import qualified Data.ByteString as SB
 import qualified Data.ByteString.Lazy as LB
+import qualified Data.ByteString.UTF8 as UTF8
+import qualified Data.List.NonEmpty as NE
+import qualified Data.Map.Strict as Map
+import           Data.Map (Map)
 import           Data.Semigroup ((<>))
+import           Data.String (IsString, fromString)
+import           Data.Text (Text)
+import qualified Data.Text as T
+import qualified Data.Text.Lazy as TL
+import qualified Data.Text.Lazy.IO as TL
 import qualified Data.Text.Lazy.Builder as Builder
+import qualified Data.Vector as V
 import qualified Formatting as F
+import           System.Directory (createDirectory, doesPathExist)
+import           System.FilePath ((</>))
+import           System.IO (hGetLine, hSetEcho, hFlush, stdout, stdin)
+import           Text.Printf (printf)
 import           Data.Time (UTCTime)
+#ifdef UNIX
+import           System.Posix.Files (ownerReadMode, setFileMode)
+#else
+import           System.Directory (emptyPermissions, readable, setPermissions)
+#endif
 
+import           Cardano.Binary (Annotated(..), reAnnotate, serialize')
 import qualified Cardano.Chain.Common as CC.Common
 import           Cardano.Chain.Common
 import           Cardano.Chain.Delegation hiding (Map, epoch)
 import           Cardano.Chain.Genesis
+import qualified Cardano.Chain.Genesis as CC.Genesis
 import           Cardano.Chain.Slotting (EpochNumber(..))
 import           Cardano.Chain.UTxO
+import qualified Cardano.Chain.UTxO as CC.UTxO
 import           Cardano.Crypto (SigningKey (..), ProtocolMagic, ProtocolMagicId)
 import qualified Cardano.Crypto.Hashing as Crypto
 import qualified Cardano.Crypto.Random as Crypto
 import qualified Cardano.Crypto.Signing as Crypto
+import           Cardano.Config.Partial (PartialCardanoConfiguration (..))
+import           Cardano.Config.Types (CardanoConfiguration(..))
 import           Cardano.Node.Configuration.Presets (mainnetConfiguration)
+import           Ouroboros.Consensus.Demo.Run
 import           Ouroboros.Consensus.Protocol hiding (Protocol)
+import qualified Ouroboros.Consensus.Protocol as Consensus
 import           Ouroboros.Consensus.Ledger.Byron
+import           Ouroboros.Consensus.Ledger.Byron.Config
 import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
 
-import           Cardano.CLI.Helpers
 import           Cardano.CLI.Ops
 import           Cardano.Config.CommonCLI
 import           Cardano.Common.Protocol
@@ -77,6 +105,46 @@ import           Cardano.Node.Orphans ()
 import           Cardano.Node.Configuration.Topology
 import           Cardano.Node.TxSubmission
 
+
+newtype GenesisFile =
+  GenesisFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype NewDirectory =
+  NewDirectory FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype SigningKeyFile =
+  SigningKeyFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype NewSigningKeyFile =
+  NewSigningKeyFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype VerificationKeyFile =
+  VerificationKeyFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype NewVerificationKeyFile =
+  NewVerificationKeyFile FilePath
+   deriving (Eq, Ord, Show, IsString)
+
+newtype CertificateFile =
+  CertificateFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype NewCertificateFile =
+  NewCertificateFile { nFp :: FilePath }
+  deriving (Eq, Ord, Show, IsString)
+
+newtype TxFile =
+  TxFile FilePath
+  deriving (Eq, Ord, Show, IsString)
+
+newtype NewTxFile =
+  NewTxFile FilePath
+  deriving (Eq, Ord, Show, IsString)
 
 data ClientCommand
   = Genesis
@@ -161,22 +229,23 @@ runCommand co@CLIOps{..}
          (Genesis
            (NewDirectory outDir)
            startTime
-           pParamsFile
+           protocolParametersFile
            blockCount
            protocolMagic
            giTestBalance
            giFakeAvvmBalance
            giAvvmBalanceFactor
            giSeed) = do
-  pParamsRaw <- LB.readFile pParamsFile
-
-  pParams <- eitherThrow
-               (ProtocolParametersParseFailed pParamsFile)
-               (canonicalDecodePretty pParamsRaw)
+  protoParamsRaw <- LB.readFile protocolParametersFile
+  protocolParameters <- case canonicalDecodePretty protoParamsRaw of
+    Left e  -> throwIO $ ProtocolParametersParseFailed protocolParametersFile e
+    Right x -> pure x
 
   -- We're relying on the generator to fake AVVM and delegation.
   mGenesisDlg <- runExceptT $ mkGenesisDelegation []
-  genesisDelegation <- eitherThrow DelegationError mGenesisDlg
+  genesisDelegation <- case mGenesisDlg of
+    Left e  -> throwIO $ DelegationError e
+    Right x -> pure x
 
   seed <- case giSeed of
     Nothing -> Crypto.runSecureRandom . Crypto.randomNumber $ shiftL 1 32
@@ -187,7 +256,7 @@ runCommand co@CLIOps{..}
         mkGenesisSpec
         genesisAvvmBalances -- :: !GenesisAvvmBalances
         genesisDelegation   -- :: !GenesisDelegation
-        pParams             -- :: !ProtocolParameters
+        protocolParameters  -- :: !ProtocolParameters
         blockCount          -- :: !BlockCount
         protocolMagic       -- :: !ProtocolMagic
         genesisInitializer  -- :: !GenesisInitializer
@@ -201,24 +270,32 @@ runCommand co@CLIOps{..}
       giUseHeavyDlg =
         True                -- Not using delegate keys unsupported.
 
-  genesisSpec <- eitherThrow GenesisSpecError mGenesisSpec
+  genesisSpec <- case mGenesisSpec of
+    Left e  -> throwIO $ GenesisSpecError e
+    Right x -> pure x
 
   mGData <- runExceptT $ generateGenesisData startTime genesisSpec
-
-  (genesisData, generatedSecrets) <- eitherThrow GenesisGenerationError mGData
+  (genesisData, generatedSecrets) <- case mGData of
+    Left e  -> throwIO $ GenesisGenerationError e
+    Right x -> pure x
 
   dumpGenesis co outDir genesisData generatedSecrets
 
-runCommand co@CLIOps{..} (PrettySigningKeyPublic skF) = do
-  sKey <- readSigningKey co skF
-  putTextLn $ prettySigningKeyPub sKey
+runCommand co@CLIOps{..} (PrettySigningKeyPublic skF) =
+  putStrLn =<< T.unpack
+             . prettySigningKeyPub
+             <$> readSigningKey co skF
 
-runCommand co (MigrateDelegateKeyFrom protocol newKeyFP oldKey) = do
-  -- Protocol specific operations
-  operations <- decideCLIOps protocol
-  sKey <- readSigningKey operations oldKey
-  newKey <- coSerialiseDelegateKey co sKey
-  LB.writeFile (nSkFp newKeyFP) newKey
+runCommand co (MigrateDelegateKeyFrom
+                  fromVer
+                  (NewSigningKeyFile newKey)
+                  oldKey) =
+        LB.writeFile newKey
+    =<< coSerialiseDelegateKey co
+    =<< flip readSigningKey oldKey
+    =<< fromCO
+  where
+    fromCO = decideCLIOps fromVer
 
 runCommand co (DumpHardcodedGenesis (NewDirectory dir)) =
   dumpGenesis co dir
@@ -234,28 +311,29 @@ runCommand CLIOps{..} (PrintGenesisHash (GenesisFile genesis)) = do
                $ snd x
 
 runCommand co@CLIOps{..} (PrintSigningKeyAddress netMagic skF) =
-  putTextLn . prettyAddress
-            . CC.Common.makeVerKeyAddress netMagic
-            . Crypto.toVerification
-            =<< readSigningKey co skF
+  putStrLn . T.unpack . prettyAddress
+           . CC.Common.makeVerKeyAddress netMagic
+           . Crypto.toVerification
+           =<< readSigningKey co skF
 
-runCommand CLIOps{ coSerialiseDelegateKey } (Keygen skF disablePassword) = do
+runCommand CLIOps{..}
+           (Keygen (NewSigningKeyFile skF) disablePassword) = do
+
   passph <- if disablePassword
             then pure Crypto.emptyPassphrase
             else readPassword $
-                 "Enter password to encrypt '" <> (nSkFp skF) <> "': "
+                 "Enter password to encrypt '" <> skF <> "': "
 
   (_vk, esk) <- Crypto.runSecureRandom $ Crypto.safeKeyGen passph
 
-  ensureNewFileLBS (nSkFp skF)
+  ensureNewFileLBS skF
     =<< (coSerialiseDelegateKey $ SigningKey $ Crypto.eskPayload esk)
 
-runCommand co (ToVerification skF (NewVerificationKeyFile vkF)) = do
-  sKey <- readSigningKey co skF
+runCommand co (ToVerification
+                skF (NewVerificationKeyFile vkF)) = do
   ensureNewFileText vkF
-    . Builder.toLazyText
-    . Crypto.formatFullVerificationKey
-    $ Crypto.toVerification sKey
+    . Builder.toLazyText . Crypto.formatFullVerificationKey . Crypto.toVerification
+    =<< readSigningKey co skF
 
 runCommand co (IssueDelegationCertificate pM epoch skF vkF cFp) = do
   sk <- readSigningKey co skF
@@ -265,14 +343,38 @@ runCommand co (IssueDelegationCertificate pM epoch skF vkF cFp) = do
   let cert = signCertificate pM vk epoch signer
   ensureNewFileLBS (nFp cFp) =<< (coSerialiseDelegationCert co $ cert)
 
-runCommand CLIOps{..} (CheckDelegation magic certFp issuerVF delegateVF) = do
+runCommand CLIOps{..}
+           (CheckDelegation magic
+            (CertificateFile certF)
+            issuerVF
+            delegateVF) = do
   issuerVK'   <- readVerificationKey issuerVF
   delegateVK' <- readVerificationKey delegateVF
-  certBS      <- LB.readFile (ceFp certFp)
-  cert :: Certificate <- eitherThrow
-                           (DlgCertificateDeserialisationFailed (ceFp certFp))
-                           (canonicalDecodePretty certBS)
-  certificateValidation magic cert delegateVK' issuerVK' certFp
+  certBS      <- LB.readFile certF
+  cert :: Certificate <- case canonicalDecodePretty certBS of
+    Left e  -> throwIO $ DlgCertificateDeserialisationFailed certF e
+    Right x -> pure x
+
+  let magic' = Annotated magic (serialize' magic)
+      epoch  = unAnnotated $ aEpoch cert
+      cert'  = cert { aEpoch = Annotated epoch (serialize' epoch) }
+      vk    :: forall r. F.Format r (Crypto.VerificationKey -> r)
+      vk     = Crypto.fullVerificationKeyF
+      f     :: forall a. F.Format Text a -> a
+      f      = F.sformat
+      issues =
+        [ f("Certificate does not have a valid signature.")
+        | not (isValid magic' cert') ] <>
+
+        [ f("Certificate issuer ".vk." doesn't match expected: ".vk)
+          (issuerVK   cert)   issuerVK'
+        |  issuerVK   cert /= issuerVK' ] <>
+
+        [ f("Certificate delegate ".vk." doesn't match expected: ".vk)
+          (delegateVK cert)   delegateVK'
+        |  delegateVK cert /= delegateVK' ]
+  unless (null issues) $
+    throwIO $ CertificateValidationErrors certF issues
 
 runCommand co (SubmitTx stTopology (TxFile stTx) stCommon) = do
   withRealPBFT co mainnetConfiguration stCommon $
@@ -305,3 +407,200 @@ runCommand co@CLIOps{..}
       putStrLn $ "genesis protocol magic:  " <> show (configProtocolMagicId gc)
       putStrLn $ "transaction hash (TxId): " <> show byronTxId
       ensureNewFileLBS ctTx (serialise tx)
+
+{-------------------------------------------------------------------------------
+  Supporting functions
+-------------------------------------------------------------------------------}
+
+-- | Perform an action that expects ProtocolInfo for Byron/PBFT,
+--   with attendant configuration.
+withRealPBFT
+  :: CLIOps IO
+  -> PartialCardanoConfiguration
+  -> CommonCLI
+  -> (RunDemo (ByronBlockOrEBB ByronConfig)
+      => CardanoConfiguration
+      -> Consensus.Protocol (ByronBlockOrEBB ByronConfig)
+      -> IO ())
+  -> IO ()
+withRealPBFT CLIOps{coProtocol} pcc common action = do
+  cc <- mkConfiguration pcc common
+  SomeProtocol p <- fromProtocol cc coProtocol
+  case p of
+    proto@ProtocolRealPBFT{} -> do
+      action cc proto
+    _ -> throwIO $ ProtocolNotSupported coProtocol
+
+signTxId :: ProtocolMagicId -> SigningKey -> TxId -> CC.UTxO.TxInWitness
+signTxId pmid sk txid = CC.UTxO.VKWitness
+  (Crypto.toVerification sk)
+  (Crypto.sign
+    pmid
+    Crypto.SignTx
+    sk
+    (CC.UTxO.TxSigData txid))
+
+txSpendUTxOByronPBFT
+  :: CC.Genesis.Config
+  -> SigningKey
+  -> NonEmpty TxIn
+  -> NonEmpty TxOut
+  -> GenTx (ByronBlockOrEBB ByronConfig)
+txSpendUTxOByronPBFT gc sk ins outs =
+  let txattrs  = mkAttributes ()
+      tx       = UnsafeTx ins outs txattrs
+      wit      = signTxId (configProtocolMagicId gc) sk (Crypto.hash tx)
+      ATxAux atx awit =
+        mkTxAux tx . V.fromList . take (NE.length ins) $ repeat wit
+  in mkByronTx $ ATxAux (reAnnotate atx) (reAnnotate awit)
+
+txSpendGenesisUTxOByronPBFT
+  :: CC.Genesis.Config
+  -> SigningKey
+  -> Address
+  -> NonEmpty TxOut
+  -> GenTx (ByronBlockOrEBB ByronConfig)
+txSpendGenesisUTxOByronPBFT gc sk genAddr outs =
+  let txattrs  = mkAttributes ()
+      tx       = UnsafeTx (pure txIn) outs txattrs
+      txIn    :: CC.UTxO.TxIn
+      txIn     = genesisUTxOTxIn gc sk genAddr
+      wit      = signTxId (configProtocolMagicId gc) sk (Crypto.hash tx) 
+      ATxAux atx awit = mkTxAux tx . V.fromList . pure $ wit
+  in mkByronTx $ ATxAux (reAnnotate atx) (reAnnotate awit)
+
+-- | Given a genesis, and a pair of signing key and address, reconstruct a TxIn
+--   corresponding to the genesis UTxO entry.
+genesisUTxOTxIn :: CC.Genesis.Config -> SigningKey -> Address -> CC.UTxO.TxIn
+genesisUTxOTxIn gc genSk genAddr =
+  let vk         = Crypto.toVerification genSk
+      handleMissingAddr :: Maybe CC.UTxO.TxIn -> CC.UTxO.TxIn
+      handleMissingAddr  = fromMaybe . error
+        $  "\nGenesis UTxO has no address\n"
+        <> (T.unpack $ prettyAddress genAddr)
+        <> "\n\nIt has the following, though:\n\n"
+        <> Cardano.Prelude.concat (T.unpack . prettyAddress <$> Map.keys initialUtxo)
+
+      initialUtxo :: Map Address (CC.UTxO.TxIn, CC.UTxO.TxOut)
+      initialUtxo =
+            Map.fromList
+          . mapMaybe (\(inp, out) -> mkEntry inp genAddr <$> keyMatchesUTxO vk out)
+          . fromCompactTxInTxOutList
+          . Map.toList
+          . CC.UTxO.unUTxO
+          . CC.UTxO.genesisUtxo
+          $ gc
+        where
+          mkEntry :: CC.UTxO.TxIn
+                  -> Address
+                  -> CC.UTxO.TxOut
+                  -> (Address, (CC.UTxO.TxIn, CC.UTxO.TxOut))
+          mkEntry inp addr out = (addr, (inp, out))
+
+      keyMatchesUTxO :: Crypto.VerificationKey -> CC.UTxO.TxOut -> Maybe CC.UTxO.TxOut
+      keyMatchesUTxO key out =
+        if CC.Common.checkVerKeyAddress key (CC.UTxO.txOutAddress out)
+        then Just out else Nothing
+
+      fromCompactTxInTxOutList :: [(CC.UTxO.CompactTxIn, CC.UTxO.CompactTxOut)]
+                               -> [(CC.UTxO.TxIn, CC.UTxO.TxOut)]
+      fromCompactTxInTxOutList =
+          map (bimap CC.UTxO.fromCompactTxIn CC.UTxO.fromCompactTxOut)
+  in handleMissingAddr $ fst <$> Map.lookup genAddr initialUtxo
+
+-- TODO:  we need to support password-protected secrets.
+readSigningKey :: CLIOps IO -> SigningKeyFile -> IO SigningKey
+readSigningKey co (SigningKeyFile fp) =
+  coDeserialiseDelegateKey co fp =<< LB.readFile fp
+
+readVerificationKey :: VerificationKeyFile -> IO Crypto.VerificationKey
+readVerificationKey (VerificationKeyFile fp) = do
+  vkB <- SB.readFile fp
+  case Crypto.parseFullVerificationKey . fromString $ UTF8.toString vkB of
+    Left e -> throwIO . VerificationKeyDeserialisationFailed fp $ T.pack $ show e
+    Right x -> pure x
+
+-- TODO:  we'd be better served by a combination of a temporary file
+--        with an atomic rename.
+
+-- | Checks if a path exists and throws and error if it does.
+ensureNewFile :: (FilePath -> a -> IO ()) -> FilePath -> a -> IO ()
+ensureNewFile writer outFile blob = do
+  exists <- doesPathExist outFile
+  when exists $
+    throwIO $ OutputMustNotAlreadyExist outFile
+  writer outFile blob
+
+ensureNewFileLBS :: FilePath -> LB.ByteString -> IO ()
+ensureNewFileLBS = ensureNewFile LB.writeFile
+
+ensureNewFileText :: FilePath -> TL.Text -> IO ()
+ensureNewFileText = ensureNewFile TL.writeFile
+
+readPassword :: String -> IO Crypto.PassPhrase
+readPassword prompt = do
+  let readOne :: String -> IO String
+      readOne pr = do
+        hPutStr stdout pr >> hFlush stdout
+        hSetEcho stdout False
+        pp <- hGetLine stdin
+        hSetEcho stdout True
+        hPutStrLn stdout ("" :: String)
+        pure pp
+      loop = do
+        (v1, v2) <- (,) <$> readOne prompt <*> readOne "Repeat to validate: "
+        if v1 == v2
+          then pure v1
+          else hPutStrLn stdout ("Sorry, entered passwords don't match." :: String)
+               >> loop
+  Crypto.PassPhrase . BA.convert . UTF8.fromString <$> loop
+
+dumpGenesis :: CLIOps IO -> FilePath -> GenesisData -> GeneratedSecrets -> IO ()
+dumpGenesis CLIOps{..} outDir genesisData GeneratedSecrets{..} = do
+  exists <- doesPathExist outDir
+  if exists
+    then throwIO $ OutputMustNotAlreadyExist outDir
+    else createDirectory outDir
+
+  let genesisJSONFile = outDir <> "/genesis.json"
+  LB.writeFile genesisJSONFile =<< coSerialiseGenesis genesisData
+
+  let dlgCertMap = unGenesisDelegation $ gdHeavyDelegation genesisData
+      isCertForSK :: SigningKey -> Certificate -> Bool
+      isCertForSK sk UnsafeACertificate{..} = delegateVK == Crypto.toVerification sk
+      findDelegateCert :: SigningKey -> IO Certificate
+      findDelegateCert sk =
+        case flip find (Map.elems dlgCertMap) . isCertForSK $ sk of
+          Nothing -> throwIO $ NoGenesisDelegationForKey $ prettySigningKeyPub sk
+          Just x  -> pure x
+      wOut :: String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
+      wOut = writeSecrets outDir
+  dlgCerts <- mapM findDelegateCert gsRichSecrets
+
+  wOut "genesis-keys"    "key"  coSerialiseGenesisKey     gsDlgIssuersSecrets
+  wOut "delegate-keys"   "key"  coSerialiseDelegateKey    gsRichSecrets
+  wOut "poor-keys"       "key"  coSerialisePoorKey        gsPoorSecrets
+  wOut "delegation-cert" "json" coSerialiseDelegationCert dlgCerts
+  wOut "avvm-seed"       "seed" (pure . LB.fromStrict)     gsFakeAvvmSeeds
+
+prettySigningKeyPub :: SigningKey -> Text
+prettySigningKeyPub (Crypto.toVerification -> vk) = TL.toStrict
+  $  "public key hash: " <> (F.format Crypto.hashHexF . CC.Common.addressHash $ vk) <> "\n"
+  <> "     public key: " <> (Builder.toLazyText . Crypto.formatFullVerificationKey $ vk)
+
+prettyAddress :: CC.Common.Address -> Text
+prettyAddress addr = TL.toStrict
+  $  F.format CC.Common.addressF         addr <> "\n"
+  <> F.format CC.Common.addressDetailedF addr
+
+writeSecrets :: FilePath -> String -> String -> (a -> IO LB.ByteString) -> [a] -> IO ()
+writeSecrets outDir prefix suffix secretOp xs =
+  forM_ (zip xs $ [0::Int ..]) $
+  \(secret, nr)-> do
+    let filename = outDir </> prefix <> "." <> printf "%03d" nr <> "." <> suffix
+    secretOp secret >>= LB.writeFile filename
+#ifdef UNIX
+    setFileMode                      filename ownerReadMode
+#else
+    setPermissions filename (emptyPermissions {readable = True})
+#endif
